// src/services/openapi/api.ts
export interface CategoryEntity {
    /**
     * 
     * @type {string}
     * @memberof CategoryEntity
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof CategoryEntity
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof CategoryEntity
     */
    'description': string;
    /**
     * 
     * @type {boolean}
     * @memberof CategoryEntity
     */
    'isActive': boolean;
}
// ...
export interface CreateProductDto {
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'productId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'productTypeId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'categoryId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'subCategoryId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'manufacturerId': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'returnExchangePolicy': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'skuid': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'productName': string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'productDescription': string;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'offerprice': number;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'actualprice': number;
    /**
     * 
     * @type {boolean}
     * @memberof CreateProductDto
     */
    'isActive': boolean;
    /**
     * 
     * @type {CreateSareeDetailsDto}
     * @memberof CreateProductDto
     */
    'productDetails': CreateSareeDetailsDto;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'cgst': number;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'sgst': number;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'maxQuantityPerCart': number;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'promoId'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof CreateProductDto
     */
    'isTrending': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof CreateProductDto
     */
    'isBestSeller': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof CreateProductDto
     */
    'isExclusive': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof CreateProductDto
     */
    'isShippable': boolean;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'collectionId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'colorId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'occassionId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'styleId'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductDto
     */
    'printId'?: string;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'maxAllowedReturnDays': number;
    /**
     * 
     * @type {number}
     * @memberof CreateProductDto
     */
    'maxAllowedCancellationDays': number;
}
// ...
export interface CreateSareeDetailsDto {
    /**
     * 
     * @type {string}
     * @memberof CreateSareeDetailsDto
     */
    'fabricDetailsId': string;
    /**
     * 
     * @type {number}
     * @memberof CreateSareeDetailsDto
     */
    'length': number;
    /**
     * 
     * @type {number}
     * @memberof CreateSareeDetailsDto
     */
    'width': number;
    /**
     * 
     * @type {boolean}
     * @memberof CreateSareeDetailsDto
     */
    'isBlousePieceIncluded': boolean;
    /**
     * 
     * @type {string}
     * @memberof CreateSareeDetailsDto
     */
    'blouseDescription': string;
}
// ...
export interface ProductTypeEntity {
    /**
     * 
     * @type {string}
     * @memberof ProductTypeEntity
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductTypeEntity
     */
    'name': ProductTypeEntityNameEnum;
    /**
     * 
     * @type {boolean}
     * @memberof ProductTypeEntity
     */
    'isActive': boolean;
    /**
     * 
     * @type {string}
     * @memberof ProductTypeEntity
     */
    'description': string;
}

export const ProductTypeEntityNameEnum = {
    Saree: 'saree'
} as const;

export type ProductTypeEntityNameEnum = typeof ProductTypeEntityNameEnum[keyof typeof ProductTypeEntityNameEnum];
// ...
export interface SareeEntity {
    /**
     * 
     * @type {string}
     * @memberof SareeEntity
     */
    'id'?: string;
    /**
     * 
     * @type {CategoryEntity}
     * @memberof SareeEntity
     */
    'category': CategoryEntity;
    /**
     * 
     * @type {SubcategoryEntity}
     * @memberof SareeEntity
     */
    'subCategory': SubcategoryEntity;
    /**
     * 
     * @type {ProductTypeEntity}
     * @memberof SareeEntity
     */
    'productType': ProductTypeEntity;
    /**
     * 
     * @type {ManufacturerEntity}
     * @memberof SareeEntity
     */
    'manufacturer': ManufacturerEntity;
    /**
     * 
     * @type {string}
     * @memberof SareeEntity
     */
    'skuid': string;
    /**
     * 
     * @type {string}
     * @memberof SareeEntity
     */
    'productName': string;
    /**
     * 
     * @type {string}
     * @memberof SareeEntity
     */
    'productDescription': string;
    /**
     * 
     * @type {number}
     * @memberof SareeEntity
     */
    'averageReview': number;
    /**
     * 
     * @type {number}
     * @memberof SareeEntity
     */
    'numberOfReviews': number;
    /**
     * 
     * @type {number}
     * @memberof SareeEntity
     */
    'offerprice': number;
    /**
     * 
     * @type {number}
     * @memberof SareeEntity
     */
    'actualprice': number;
    /**
     * 
     * @type {number}
     * @memberof SareeEntity
     */
    'available_qty': number;
    /**
     * 
     * @type {boolean}
     * @memberof SareeEntity
     */
    'isActive': boolean;
    /**
     * 
     * @type {string}
     * @memberof SareeEntity
     */
    'returnExchangePolicy': string;
    /**
     * 
     * @type {number}
     * @memberof SareeEntity
     */
    'cgst': number;
    /**
     * 
     * @type {number}
     * @memberof SareeEntity
     */
    'sgst': number;
    /**
     * 
     * @type {number}
     * @memberof SareeEntity
     */
    'maxQuantityPerCart': number;
    /**
     * 
     * @type {PromoDetailsEntity}
     * @memberof SareeEntity
     */
    'promoDetails': PromoDetailsEntity;
    /**
     * 
     * @type {boolean}
     * @memberof SareeEntity
     */
    'isNew': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof SareeEntity
     */
    'isTrending': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof SareeEntity
     */
    'isBestSeller': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof SareeEntity
     */
    'isExclusive': boolean;
    /**
     * 
     * @type {boolean}
     * @memberof SareeEntity
     */
    'isShippable': boolean;
    /**
     * 
     * @type {number}
     * @memberof SareeEntity
     */
    'maxAllowedReturnDays': number;
    /**
     * 
     * @type {number}
     * @memberof SareeEntity
     */
    'maxAllowedCancellationDays': number;
    /**
     * 
     * @type {SareeDetailsEntity}
     * @memberof SareeEntity
     */
    'sareeDetails': SareeDetailsEntity;
    /**
     * 
     * @type {Array<SareeImageEntity>}
     * @memberof SareeEntity
     */
    'productImages': Array<SareeImageEntity>;
    /**
     * 
     * @type {ProductCollectionEntity}
     * @memberof SareeEntity
     */
    'collection': ProductCollectionEntity;
    /**
     * 
     * @type {ProductColorEntity}
     * @memberof SareeEntity
     */
    'colour': ProductColorEntity;
    /**
     * 
     * @type {ProductPrintsEntity}
     * @memberof SareeEntity
     */
    'print': ProductPrintsEntity;
    /**
     * 
     * @type {ProductOccassionEntity}
     * @memberof SareeEntity
     */
    'occassion': ProductOccassionEntity;
    /**
     * 
     * @type {ProductStyleEntity}
     * @memberof SareeEntity
     */
    'style': ProductStyleEntity;
}
// ...
export interface SubcategoryEntity {
    /**
     * 
     * @type {string}
     * @memberof SubcategoryEntity
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof SubcategoryEntity
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof SubcategoryEntity
     */
    'description': string;
    /**
     * 
     * @type {boolean}
     * @memberof SubcategoryEntity
     */
    'isActive': boolean;
    /**
     * 
     * @type {CategoryEntity}
     * @memberof SubcategoryEntity
     */
    'category': CategoryEntity;
}
// ...
export class ProductApi extends BaseAPI {
    /**
     * 
     * @param {CreateProductDto} createProductDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductApi
     */
    public productControllerCreate(createProductDto: CreateProductDto, options?: RawAxiosRequestConfig) {
        return ProductApiFp(this.configuration).productControllerCreate(createProductDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {boolean} pagingRequired 
     * @param {number} page 
     * @param {number} limit 
     * @param {string} productTypeId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductApi
     */
    public productControllerFindAll(pagingRequired: boolean, page: number, limit: number, productTypeId: string, options?: RawAxiosRequestConfig) {
        return ProductApiFp(this.configuration).productControllerFindAll(pagingRequired, page, limit, productTypeId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} productTypeId 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductApi
     */
    public productControllerFindOne(productTypeId: string, id: string, options?: RawAxiosRequestConfig) {
        return ProductApiFp(this.configuration).productControllerFindOne(productTypeId, id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} productTypeId 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductApi
     */
    public productControllerRemove(productTypeId: string, id: string, options?: RawAxiosRequestConfig) {
        return ProductApiFp(this.configuration).productControllerRemove(productTypeId, id, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class ProductImageApiApi extends BaseAPI {
    /**
     * 
     * @param {string} imageDescription 
     * @param {string} productId 
     * @param {string} productTypeId 
     * @param {File} file 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductImageApiApi
     */
    public productImageControllerUploadProductImage(imageDescription: string, productId: string, productTypeId: string, file: File, options?: RawAxiosRequestConfig) {
        return ProductImageApiApiFp(this.configuration).productImageControllerUploadProductImage(imageDescription, productId, productTypeId, file, options).then((request) => request(this.axios, this.basePath));
    }
}